prog: speed up resource ctors detection
When we build a list of resource constructors we over and over iterate through
all types in a syscall to find resource types. Speed it up by iterating only
once to build a list of resource types each syscall can produce and then reuse
it. This significantly speeds up syz-exeprog startup time on Raspberry Pi Zero.